Refactoring (2/3 of the way or so)
Began refactoring for maintainability and readability through the
AlarmsTimers class for now (the majority of breaking into smaller
methods has been completed).  Be sure to run this code past the geeks on
Utopia to see how much better it might be than the preceeding crap that
was all hashed up by them.