Fixed timer interval/other minor bugfixes
Fixed the setting of the timer interval so that it is not set as an
alarm.  Notice that checkAlarmDay() is not correctly resetting the
DateTime 'target' object in the AlarmsTimers class properly, resulting
in negative intervals depending on how old the 'target' is.  Haven't
quite figured out that bug yet, probably need to rewrite
checkAlarmDay().