Working on _Tick modularization and fixes
_Tick has some pretty major bugs that need to be fixed due to improper
setting and unsetting of when it's enabled and such; also if one alarm
is going off and it is acknowledged then it doesn't allow concurrent
timers or alarms to continue, etc.  I'm using the time going through
this to create more modular and reusable code in the meantime in order
to make future debugging simpler and make the whole application easier
to understand as per advice from seasoned developers.