Modularization on _Tick ringRingNeo()
Worked more on modularizing _Tick; got ringRingNeo() written from the
original code to play the music/beep that was duplicated in both the
alarms & timers section.  An index out of range exception is now found
in the alarm ringing code at least for alarms, though, so debugging is
needed as usual.